Fix an issue I ran into with the latest kernel
For some reason I started getting kernel panics on boot which I
attributed to a kernel update. Disabling the gobook-xr1-sound init
script fixed the issue and sound kept working. I experimented with the
blacklist file also and found that only the realtek driver needs to be
blacklisted for sound to work.